Fix part of #10415: Refactor classifier validators (#11116)
* Refactor classifier validators

* Refactor classifier validators

* Refactor classifier validators

* refactor classifier validators

* refactor classifier validators

* Refactor classifier validators
